Add initial support -fcoverage-mapping support
This is enought make it work up until llvm-cov tries to read the
named data sections in the binary and can't find them.  For this
final part to work we probably need to switch the object format to
using multiple code and data sections:
https://github.com/WebAssembly/tool-conventions/issues/138

Not sure if its worth submitting this part in isolation without
a fully working solution?

See #13046